requirements.txt

streamlit
pdfminer.six
python-docx
spacy
sentence-transformers
scikit-learn
